+++ NEWS.md
@@ -53,6 +53,11 @@
 
 - Fixed an issue in `layer_input()` where passing a tensorflow `DType` objects to argument `dtype` would throw an error.
 
+- Fixed an issue in `compile()` where passing an R function via an in-line 
+  call would result in an error from subsequent `fit()` calls.
+  (e.g., `compile(loss = function(y_true, y_pred) my_loss(y_true, y_pred))` 
+  now succeeds)
+
 - `clone_model()` gains a `clone_function` argument that allows you to customize each layer as it is cloned.
 
 - Existing layer instances gain the ability to be added to sequential models via a call. E.g.: